\name{ExtractFromApopDataList}
\alias{ExtractFromApopDataList}
\title{Extract elements from an apop_data list...}
\usage{ExtractFromApopDataList(mlist, what)}
\description{Extract elements from an apop_data list
generated from a call to RCheckConsistency}
\value{the element of the list asked for}
\author{Rolando Rodriguez \email{rolando.a.rodriguez@census.gov}}
\arguments{\item{mlist}{a list of apop_data elements from RCheckConsistency}
\item{what}{a character string of what to get}}
